adding dev and bats using binary

+++ Makefile
@@ -1,3 +1,5 @@
+DIRS = config crypto document entity fs index node x509
+
 default: get-deps build test
 
 get-deps:
@@ -8,8 +10,15 @@ build:
 install:
 	install -m 0755 pki.io /usr/local/bin
 test:
-	gom exec bats bats_tests
+	bats bats_tests
 clean:
-	rm pki.io
+	test ! -d _vendor || rm -rf _vendor/*
+	test ! -e pki.io || rm pki.io
+
+dev: clean get-deps
+	test -d _vendor/src/github.com/pki-io/pki.io  && \
+	rm -rf _vendor/src/github.com/pki-io/pki.io/* && \
+	for d in $(DIRS); do (cd _vendor/src/github.com/pki-io/pki.io && ln -s ../../../../../../pki.io/$$d .); done && \
+	rm -rf _vendor/pkg
 
 all: get-deps build test install
